Add SDR/audio integration (Phase 4) and compile 69 external apps into binary (Phase 5)
Phase 4 — SDR + Audio + Baseband:
- ARM SIMD intrinsics shim (~35 intrinsics as pure C++ scalar ops)
- All 48 baseband processors compile on x86_64 via processor factory
- SoapySDR abstraction layer for HackRF/RTL-SDR/LimeSDR support
- SDL2 audio output with lock-free SPSC ring buffer (48kHz stereo)
- Baseband thread: SoapySDR IQ → processor → audio pipeline
- Radio stubs wired to SoapySDR, m4_init() launches baseband thread

Phase 5 — External Apps:
- All 69 external apps compiled directly into the binary (no .ppma files needed)
- External app registry collects application_information_t structs at link time
- Menu loader uses registry under LINUX_SHIM (direct function call, no SD scan)
- Shim fixes: Thread::p_refs, chRegSetThreadName, FATFS::drv/fatbase,
  disk_write stubs, std::min<uint32_t> for size_t width mismatch

+++ firmware/baseband/proc_flash_utility.cpp
@@ -100,7 +100,7 @@ void write_firmware(FIL* firmware_file) {
         if (page_index % 32 == 0)
             palTogglePad(LED_PORT, LEDTX_PAD);
 
-        size_t bytes_read;
+        UINT bytes_read;
         if (f_read(firmware_file, data_buffer, PAGE_LEN, &bytes_read) != FR_OK) chDbgPanic("no data");
 
         if (bytes_read > 0)
